dbadpter以外のテストまで実装

+++ api/src/activate.ts
@@ -4,16 +4,39 @@ import { APIGatewayProxyEventQueryStringParameters, APIGatewayProxyResultV2 } fr
 import dbadapter from "./dbadapter";
 const logger = common.getLogger();
 export default async(params: APIGatewayProxyEventQueryStringParameters): Promise<APIGatewayProxyResultV2>=>{
-    if(params.code) {
+    if(params.code && params.user_id) {
         const activationCode = await dbadapter.getActivationCode(params.code);
         if(activationCode) {
+            if(!await dbadapter.setSharedIdToTrashSchedule(params.user_id, activationCode.shared_id)) {
+                return {
+                    statusCode: 500
+                }
+            }
             logger.debug(`Get Activation Code -> ${JSON.stringify(activationCode)}`);
-            const syncResult = sync({ user_id: activationCode.user_id });
+            const sharedSchedule = await dbadapter.getSharedScheduleBySharedId(activationCode.shared_id);
+            if(sharedSchedule === null) {
+                return {
+                    statusCode: 500
+                }
+            }
+            const updateTrashScheduleResult = dbadapter.updateTrashSchedule(params.user_id, sharedSchedule.description, sharedSchedule.timestamp);
             // 使い終わったActivationCodeを削除する
-            const deleteResult = await dbadapter.deleteActivationCode(activationCode.code);
-            const response = await Promise.all([syncResult, deleteResult]);
+            const deleteResult = dbadapter.deleteActivationCode(activationCode.code);
+            const response = await Promise.all([updateTrashScheduleResult, deleteResult]);
             logger.debug(JSON.stringify(response));
-            return response[0];
+
+            if(response[0]) {
+                return {
+                    statusCode: 200,
+                    body: JSON.stringify({
+                        description: sharedSchedule.description
+                    })
+                };
+            } else {
+                return {
+                    statusCode: 500
+                }
+            }
         } else {
             logger.error(`Activation Code Not Found: ${params.code}`)
             return {
